Update bra2 branch

+++ Backend/src/main/java/com/example/mini_projet/Municipality/MunicipalityRepository.java
@@ -9,11 +9,11 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface MunicipalityRepository extends JpaRepository<Municipality , Long> {
+public interface MunicipalityRepository extends JpaRepository<Municipality , Integer> {
 
 
 
-    Optional<Municipality> findById(Long idMun);
+    Optional<Municipality> findById(Integer idMun);
 
 
 
@@ -22,7 +22,7 @@ public interface MunicipalityRepository extends JpaRepository<Municipality , Lon
 
 
     @Query("SELECT m FROM Municipality m  WHERE m.idMun = ?1")
-    Municipality findByCodeMuni(Long idMun);
+    Municipality findByCodeMuni(Integer idMun);
 
 
 }
