+++ Backend/src/main/java/com/example/mini_projet/TEST/SubAdminController.java
@@ -41,7 +41,7 @@ public class SubAdminController {
 
 
     @PostMapping("/add/municipality/{id}")
-    public ResponseEntity<String> addStatistic(@RequestBody Statistics s , @PathVariable("id") Long idMun){
+    public ResponseEntity<String> addStatistic(@RequestBody Statistics s , @PathVariable("id") Integer idMun){
 
         boolean test = statisticsService.insertStat(s , idMun);
         if(test){
